ecom-airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-snowflake>=6.9.0
astronomer-cosmos>=1.13.0
dbt-core>=1.11.3
dbt-snowflake>=1.11.1
apache-airflow-providers-google==19.5.0